Move RuleFilter from App\Rule to App\Analyzer namespace
Moving RuleFilter to the Analyzer namespace is cleaner than excluding it
from service container configuration. Since RuleFilter is a utility class
used by RstAnalyzer (not a Rule implementation), it makes more semantic
sense to have it in the Analyzer namespace.

Changes:
- Move src/Rule/RuleFilter.php to src/Analyzer/RuleFilter.php
- Move tests/Rule/RuleFilterTest.php to tests/Analyzer/RuleFilterTest.php
- Update namespace from App\Rule to App\Analyzer
- Remove exclude directive from services.php
- Remove unused import from services.php
